+++ test/core_tests/value_reader_test.py
@@ -17,7 +17,7 @@ class TestValueReader:
     async def test_value_reader_read_success(self):
         """Test value reader: successful read."""
         xknx = XKNX()
-        test_group_address = GroupAddress("0/0/0")
+        test_group_address = GroupAddress("0/0/1")
         response_telegram = Telegram(
             destination_address=test_group_address,
             direction=TelegramDirection.INCOMING,
@@ -43,7 +43,7 @@ class TestValueReader:
     async def test_value_reader_read_timeout(self, logger_warning_mock):
         """Test value reader: read timeout."""
         xknx = XKNX()
-        value_reader = ValueReader(xknx, GroupAddress("0/0/0"))
+        value_reader = ValueReader(xknx, GroupAddress("0/0/1"))
         value_reader.response_received_event.wait = MagicMock(
             side_effect=asyncio.TimeoutError()
         )
@@ -56,7 +56,7 @@ class TestValueReader:
         logger_warning_mock.assert_called_once_with(
             "Error: KNX bus did not respond in time (%s secs) to GroupValueRead request for: %s",
             2.0,
-            GroupAddress("0/0/0"),
+            GroupAddress("0/0/1"),
         )
         # Callback was removed again
         assert not xknx.telegram_queue.telegram_received_cbs
@@ -68,7 +68,7 @@ class TestValueReader:
     async def test_value_reader_read_cancelled(self):
         """Test value reader: read cancelled."""
         xknx = XKNX()
-        value_reader = ValueReader(xknx, GroupAddress("0/0/0"))
+        value_reader = ValueReader(xknx, GroupAddress("0/0/1"))
         value_reader.response_received_event.wait = MagicMock(
             side_effect=asyncio.CancelledError()
         )
@@ -85,19 +85,19 @@ class TestValueReader:
     async def test_value_reader_send_group_read(self):
         """Test value reader: send_group_read."""
         xknx = XKNX()
-        value_reader = ValueReader(xknx, GroupAddress("0/0/0"))
+        value_reader = ValueReader(xknx, GroupAddress("0/0/1"))
 
         await value_reader.send_group_read()
         assert xknx.telegrams.qsize() == 1
         telegram = xknx.telegrams.get_nowait()
         assert telegram == Telegram(
-            destination_address=GroupAddress("0/0/0"), payload=GroupValueRead()
+            destination_address=GroupAddress("0/0/1"), payload=GroupValueRead()
         )
 
     async def test_value_reader_telegram_received(self):
         """Test value reader: telegram_received."""
         xknx = XKNX()
-        test_group_address = GroupAddress("0/0/0")
+        test_group_address = GroupAddress("0/0/1")
         expected_telegram_1 = Telegram(
             destination_address=test_group_address,
             direction=TelegramDirection.INCOMING,
@@ -109,7 +109,7 @@ class TestValueReader:
             payload=GroupValueWrite(DPTBinary(1)),
         )
         telegram_wrong_address = Telegram(
-            destination_address=GroupAddress("0/0/1"),
+            destination_address=GroupAddress("0/0/2"),
             direction=TelegramDirection.INCOMING,
             payload=GroupValueResponse(DPTBinary(1)),
         )