+++ xknx/remote_value/remote_value_color_rgb.py
@@ -23,8 +23,8 @@ class RemoteValueColorRGB(RemoteValue[tuple[int, int, int]]):
     def __init__(
         self,
         xknx: XKNX,
-        group_address: GroupAddressesType | None = None,
-        group_address_state: GroupAddressesType | None = None,
+        group_address: GroupAddressesType = None,
+        group_address_state: GroupAddressesType = None,
         sync_state: bool | int | float | str = True,
         device_name: str | None = None,
         feature_name: str = "Color RGB",