+++ xknx/remote_value/remote_value_control.py
@@ -23,8 +23,8 @@ class RemoteValueControl(RemoteValue[Any]):
     def __init__(
         self,
         xknx: XKNX,
-        group_address: GroupAddressesType | None = None,
-        group_address_state: GroupAddressesType | None = None,
+        group_address: GroupAddressesType = None,
+        group_address_state: GroupAddressesType = None,
         sync_state: bool | int | float | str = True,
         value_type: str | None = None,
         device_name: str | None = None,