+++ xknx/remote_value/remote_value_datetime.py
@@ -32,8 +32,8 @@ class RemoteValueDateTime(RemoteValue[time.struct_time]):
     def __init__(
         self,
         xknx: XKNX,
-        group_address: GroupAddressesType | None = None,
-        group_address_state: GroupAddressesType | None = None,
+        group_address: GroupAddressesType = None,
+        group_address_state: GroupAddressesType = None,
         sync_state: bool | int | float | str = True,
         value_type: str = "time",
         device_name: str | None = None,