+++ xknx/remote_value/remote_value_switch.py
@@ -22,8 +22,8 @@ class RemoteValueSwitch(RemoteValue[bool]):
     def __init__(
         self,
         xknx: XKNX,
-        group_address: GroupAddressesType | None = None,
-        group_address_state: GroupAddressesType | None = None,
+        group_address: GroupAddressesType = None,
+        group_address_state: GroupAddressesType = None,
         sync_state: bool | int | float | str = True,
         device_name: str | None = None,
         feature_name: str = "State",